Ported `hanami server` to component based architecture
