+++ lib/hanami/commands/apps.rb
@@ -1,0 +1,4 @@
+module Hanami
+  module Components
+  end
+end
